requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.14.2
better-ffmpeg-progress==4.0.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
cython==3.1.5
dateparser==1.1.8
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
flask==3.1.2
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
g2pkk==0.1.2
gitdb==4.0.12
gitpython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.32.0
google-genai==1.46.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflect==7.5.0
iprogress==0.4
ipykernel==7.0.1
ipython==8.37.0
isoduration==20.11.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.10
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
kiwisolver==1.4.9
langchain==1.0.2
langchain-core==1.0.0
langchain-google-genai==3.0.0
langcodes==3.5.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
language-data==1.3.0
lark==1.3.0
lazy-loader==0.4
librosa==0.10.0
llvmlite==0.45.1
marisa-trie==1.3.1
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
narwhals==2.10.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.9.2
notebook==7.4.7
notebook-shim==0.2.4
num2words==0.5.14
numba==0.62.1
numpy==2.2.6
openai-whisper==20250625
opencv-python==4.12.0.88
orjson==3.11.3
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
pip @ file:///home/task_176054724192090/conda-bld/pip_1760547668629/work
platformdirs==4.5.0
pooch==1.8.2
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
proglog==0.1.12
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.1
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydeck==0.9.1
pydub==0.25.1
pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
pypinyin==0.55.0
pysbd==0.3.4
pysrt==1.1.2
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
pytube==15.0.0
pytz==2025.2
pywinpty==3.0.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
send2trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.4.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
streamlit==1.51.0
sudachidict-core==20250825
sudachipy==0.6.10
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.3.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.2
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.33.0
tts==0.22.0
typeguard==4.4.4
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
unidecode==1.4.0
update-checker==0.18.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.9.5
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.10.22
zstandard==0.25.0
